backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
